Python/LangChain/requirements.txt
# =========================
# LangChain Core
# =========================
langchain
langchain-core
langchain-community

# =========================
# OpenAI Integration
# =========================
langchain-openai
openai

# =========================
# Anthropic Integration
# =========================
langchain-anthropic

# =========================
# Google Gemini Integration
# =========================
langchain-google-genai
google-generativeai

# =========================
# Hugging Face Integration
# =========================
langchain-huggingface
transformers>=4.42
huggingface-hub
accelerate
safetensors
sentencepiece
sentence_transformers

# =========================
# Environment Variable Management
# =========================
python-dotenv

# =========================
# Machine Learning Utilities
# =========================
numpy
scikit-learn

# =========================
# PyTorch (CPU Build)
# =========================
#torch==2.2.2

# =========================
# PyTorch (GPU Build)
# =========================

# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# =========================
# UI Interface
# =========================
streamlit
grandalf
pdfplumber
pypdf
beautifulsoup4 
chromadb
tiktoken
faiss-gpu
hnswlib